frontend: Rename auth directory
